Cambia el tipo de dato de 'quantity' a Integer en los modelos de order_items e inventory_movements, actualizando esquemas y lógica relacionada. Mejora la gestión de cantidades en OrderNew y ajusta estilos en OrderDetail.

+++ Backend/app/models/inventory_movement.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Numeric, ForeignKey
+from sqlalchemy import Column, String, ForeignKey, Integer
 from sqlalchemy.dialects.postgresql import UUID
 import uuid
 from app.database import Base
@@ -12,5 +12,5 @@ class InventoryMovement(Base, AuditMixin):
     material_id = Column(UUID(as_uuid=True), ForeignKey("materials.id"), nullable=False)
     order_id = Column(UUID(as_uuid=True), ForeignKey("orders.id"), nullable=True)  # si el movimiento vino de un pedido
     movement_type = Column(String, nullable=False)  # entrada | salida | ajuste
-    quantity = Column(Numeric(10, 2), nullable=False)  # positivo en entrada/salida; puede ser +/- en ajuste
+    quantity = Column(Integer, nullable=False)  # positivo en entrada/salida; puede ser +/- en ajuste
     reason = Column(String, nullable=True)
